+++ Runtime/Library/Value/ULongExtensions.cs
@@ -250,7 +250,7 @@ namespace BitCore
 			ulong value = 0;
 			for (int i = 0; i < 64; i++)
 			{
-				value = (value << 1) | (ulong)(data[readIndex + i] - '0');
+				value = (value << 1) | (uint)(data[readIndex + i] - '0');
 			}
 			return value;
 		}
